web/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pyzmq==26.0.0
msgpack==1.0.7
